+++ src/main/java/it/tredi/msa/mailboxmanager/docway/Docway4MailboxManager.java
@@ -7,7 +7,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import it.tredi.utils.properties.PropertiesReader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.dom4j.Attribute;
@@ -783,8 +782,8 @@ public class Docway4MailboxManager extends DocwayMailboxManager {
 	private void sendNotificationEmail(MailSender mailSender, String senderAddress, String senderPersonal, String matricola, boolean isRPA, DocwayDocument doc, Document savedDocument, String body, String codAmmAooDestinatario, Set<String> notifiedAddresses) {
 		Docway4MailboxConfiguration conf = (Docway4MailboxConfiguration)getConfiguration();
 		try {
-			boolean nrecordOnSubject = conf.isAddNrecordOnNotification();
-			String subject = Docway4NotificationEmailsUtils.getSubjectForEmail(isRPA?"RPA":"CC", savedDocument, nrecordOnSubject ? matricola : null);
+			boolean matricolaOnSubject = conf.isAddMatricolaOnNotification();
+			String subject = Docway4NotificationEmailsUtils.getSubjectForEmail(isRPA?"RPA":"CC", savedDocument, matricolaOnSubject ? matricola : null);
 			String destEmail = getEmailWithMatricola(matricola, codAmmAooDestinatario);
 			String []destinatari = destEmail.split(",");
 			for (String dest:destinatari) {